Extract `spec/unit.ts`
